scrapper/requirements.txt
beautifulsoup4==4.11.2
bs4==0.0.1
lxml==4.9.2
python-dotenv==1.0.0
